Grades: Clean up tests EDUCATOR-1404
